+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.9.0] - 2024-02-06
+
 ### Added
 
 - Add mustache templating in loader and batch names [#320](https://github.com/nbbrd/java-service-util/issues/320)
@@ -160,7 +162,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/nbbrd/java-service-util/compare/v1.8.0...HEAD
+[Unreleased]: https://github.com/nbbrd/java-service-util/compare/v1.9.0...HEAD
+[1.9.0]: https://github.com/nbbrd/java-service-util/compare/v1.8.0...v1.9.0
 [1.8.0]: https://github.com/nbbrd/java-service-util/compare/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/nbbrd/java-service-util/compare/v1.6.1...v1.7.0
 [1.6.1]: https://github.com/nbbrd/java-service-util/compare/v1.6.0...v1.6.1